requirements.txt
asgiref==3.2.3
captcha==0.3
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.10
django-cors-headers==3.2.1
django-crispy-forms==1.8.1
django-formtools==2.2
django-import-export==2.0.1
django-ranged-response==0.2.0
django-reversion==3.0.5
django-simple-captcha==0.5.12
django-six==1.0.4
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.12.3
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
Pillow==6.2.1
PyMySQL==0.9.3
pytz==2019.3
PyYAML==5.3
six==1.12.0
sqlparse==0.3.0
tablib==0.14.0
xadmin-py3==2.0.4
xlrd==1.2.0
xlwt==1.3.0
